requirements.txt
supabase==2.3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
alembic==1.13.1
aiohttp==3.12.15